feat: Added ability to disable writing to the DB.
If in read only mode, then a commit hook is added to the connection to
reload the page.

The ideal setup now is to have one "leader" and many followers.
The leader has the content directory, and watches for changes,
whereas the followers only have the cache and db.

If you're using emphasize for local development, then you only need one
instance in read write mode.